+++ components/supervisor/frontend/src/ide/ide-web-socket.ts
@@ -8,7 +8,7 @@ import { serverUrl } from "../shared/urls";
 import { metricsReporter } from "./ide-metrics-service-client";
 import ReconnectingWebSocket from "reconnecting-websocket";
 import { Disposable } from "@gitpod/gitpod-protocol/lib/util/disposable";
-import { getUrlProvider } from "@gitpod/gitpod-protocol/lib/messaging/browser/connection";
+import { getUrlProvider } from "@gitpod/gitpod-protocol/lib/messaging/browser/url-provider";
 
 let connected = false;
 const workspaceSockets = new Set<IDEWebSocket>();
